Make datetime rendering more consistent
