+++ caldir-cli/src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod date;
 pub mod tui;
